+++ src/Prelude/Backprop/Num.hs
@@ -23,6 +23,7 @@ module Prelude.Backprop.Num (
   , minimum
   , maximum
   , traverse
+  , toList
   -- * Functor and Applicative
   , fmap
   , (<$>)
@@ -209,3 +210,14 @@ realToFrac
 realToFrac = liftOp1 . op1 $ \x ->
     (P.realToFrac x, P.realToFrac)
 {-# INLINE realToFrac #-}
+
+-- | Lifted version of 'P.toList'.  Takes a 'BVar' of a 'Traversable' of
+-- items and returns a list of 'BVar's for each item.
+--
+-- @since 0.2.2.0
+toList
+    :: (Traversable t, Num a, Reifies s W)
+    => BVar s (t a)
+    -> [BVar s a]
+toList = toListOfVar P.traverse
+{-# INLINE toList #-}